Update patternrecognition.md
I thought I already fixed the issues when you first sent the emails regarding the citations but I think I might edited in the wrong place and the changes never went through. Since I was working on the new glossary words, I realized it was still the old version for this word so I updated it.